requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
blessings==1.7
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
curtsies==0.3.4
dj-database-url==0.5.0
Django==3.1.2
django-cli-g73==0.1.1
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
enquiries==0.1.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
pep8==1.7.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyfiglet==0.8.post1
pylint==2.6.0
PyMySQL==0.10.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1


